+++ src/views/settings/components/JobQueue.vue
@@ -24,7 +24,7 @@
       <setting :setting-group="hackneyPools" :data="hackneyPoolsData"/>
     </el-form>
     <div class="submit-button-container">
-      <el-button class="submit-button" type="primary" @click="onSubmit">Submit</el-button>
+      <el-button class="submit-button" type="primary" @click="onSubmit">{{ $t('settings.submit') }}</el-button>
     </div>
   </div>
 </template>