+++ src/views/settings/components/Other.vue
@@ -26,7 +26,7 @@
       <setting :setting-group="castAndValidate" :data="castAndValidateData"/>
     </el-form>
     <div class="submit-button-container">
-      <el-button class="submit-button" type="primary" @click="onSubmit">Submit</el-button>
+      <el-button class="submit-button" type="primary" @click="onSubmit">{{ $t('settings.submit') }}</el-button>
     </div>
   </div>
 </template>